+++ nkululeko/constants.py
@@ -1,4 +1,4 @@
-VERSION = "1.8.3"
+VERSION = "1.8.4"
 SAMPLING_RATE = 16000
 COL_SEX = "gender"
 COL_AGE = "age"