+++ nkululeko/feat_extract/feats_squim.py
@@ -101,6 +101,7 @@ class SquimSet(Featureset):
         return pesq, sdr, stoi
 
     def extract_sample(self, signal, sr):
-        self.init_model()
+        if not self.model_initialized:
+            self.init_model()
         feats = self.get_embeddings(signal, sr, "no file")
         return feats