+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 Changelog
 =========
 
+Version 1.8.4 (26-08-16)
+------------------------
+* fix bug: unguarded init_model
+
 Version 1.8.3 (26-08-16)
 ------------------------
-* fux bug: best model reload
+* fix bug: best model reload
 
 Version 1.8.2 (26-08-16)
 ------------------------